Restructure to prepare for OOP
- Free `__init__.py` from a bunch of processing code, keeping concern on
  metadata only. Now all code for preferences, menus and keymaps belongs
  to `preferences.py`; `operators.py` is responsible for all operators.
- Every module is responsible for registering and unregistering its
  components from now.
- Push `bl_info` to the top of the file, so we have to ignore E402 of
  'Flake8' from `__init__.py`.
