+++ imagepaste/clipboard/windows/windows.py
